requirements.txt
Django==5.1.1
djangorestframework==3.15.2
pytz==2024.2
skyfield==1.49
